autonomous_rpa_agent/requirements.txt
langgraph
langchain-google-genai
gradio
selenium
pyautogui
pillow
mss
python-dotenv
webdriver-manager
